refactor: move gnu_basename/gnu_dirname to utils/path
