+++ python/mirage/utils/path.py
@@ -28,3 +28,34 @@ def resolve_path(path: str, cwd: str) -> str:
     if not path.startswith("/"):
         path = cwd.rstrip("/") + "/" + path
     return posixpath.normpath(path)
+
+
+def gnu_basename(path: str, suffix: str | None = None) -> str:
+    i = len(path)
+    while i > 0 and path[i - 1] == "/":
+        i -= 1
+    if i == 0:
+        return "/" if path else ""
+    j = path.rfind("/", 0, i)
+    base = path[j + 1:i]
+    if suffix and base != suffix and base.endswith(suffix):
+        base = base[:len(base) - len(suffix)]
+    return base
+
+
+def gnu_dirname(path: str) -> str:
+    if path == "":
+        return "."
+    i = len(path)
+    while i > 0 and path[i - 1] == "/":
+        i -= 1
+    if i == 0:
+        return "/"
+    j = path.rfind("/", 0, i)
+    if j == -1:
+        return "."
+    while j > 0 and path[j - 1] == "/":
+        j -= 1
+    if j == 0:
+        return "/"
+    return path[:j]
